implement map and tables with responsive height
